math: Vec3f: impl right-hand scalar arithmetic
Allow arithmetic with a vector on the left and a scalar on the right,
where the scalar gets promoted to a vector by "splatting" it across all
components.

The macro-generated division by scalar is technically inferior to a
handwritten, but I don't want to bother decomposing the macros to make
that possible yet. Maybe I'll just use `nalgebra_glm` once I'm satisfied
with this exercise.