math: Vec3f: switch to named struct
The tuple struct constructor syntax started getting a little weird when
I was thinking about `::splat()` (see future commits). Switch to a named
struct with a `::new()` function for that consistency. To make the
interface a bit more uniform, add `From` and `Index` impls, and make the
field private to avoid temptation to bypass these.

This commit is getting a little large, but I had to update the `BASIS_I`
etc definitions, and I realized in the process of doing this that I
didn't want to assume _any_ component or basis names in the core `Vec3f`
type. (It's up to newtypes to define XYZ, RGB, YUV, STP, etc.) So I
changed the names to `BASIS_0` etc, and I promoted them to associated
constants for easier reuse. This little bit of dogfooding prompted me to
make `::new()` a `const fn` as well.